Table hunt: Refactor code to enable fz_find_table_within_bounds
The existing fz_table_hunt() function searches through the page
looking to find the bounds of as many tables as it can find. It
then rewrites the page to have the table content split out into
table form.

We refactor some of the code here (no functional change) so that
we can reuse the same logic to implement a new function that
does the rewriting for a single table, whose bounds are passed
in.

Thus fz_find_table_within_bounds avoids the 'hunt' for the tables
and instead relied on the bounds passed in by the user.