Table hunt: Add bounds to the internal table hunt data.
When we are searching on an entire page, these can be set to
infinite. When we have been given a particular rectangle to look
in, then set the bounds to this rectangle.

Use these bounds to constrain vector graphics. This stops us
rejecting vectors just because they happen to overrun the edge
of our bounded regions.

This is useful both when extracting 'subtables', or when the
groundtruth bound isn't quite large enough to cover invisible
background rectangles.